fraud-detection-msc/requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Imbalanced Learning
imbalanced-learn==0.11.0

# Machine Learning
xgboost==1.7.6
lightgbm==4.0.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Jupyter
jupyter==1.0.0
ipykernel==6.25.0
notebook==7.0.2

# Experiment Tracking
mlflow==2.5.0

# API Development
fastapi==0.100.0
uvicorn[standard]==0.23.1
pydantic==2.1.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Data Download
kaggle==1.5.16

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.1

# Code Quality
black==23.7.0
flake8==6.0.0
mypy==1.4.1

# Statistical Testing
statsmodels==0.14.0
